requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-MySQL==1.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
PyMySQL==1.1.1
Werkzeug==3.1.3
Flask>=2.2.0
Flask-WTF>=1.0.0
mysql-connector-python>=8.0.0
Flask-Migrate>=4.0.0
